requirements.txt
opencv-python
mediapipe
tensorflow
streamlit
pyttsx3
numpy
pillow
scikit-learn
